initial commit for unpivoted prices data
